refactor: extract common get_s3_url_reader method
